+++ src/main/java/me/manuelp/jevsto/EventStore.java
@@ -2,24 +2,31 @@ package me.manuelp.jevsto;
 
 import fj.data.List;
 import fj.data.Option;
+import java.util.UUID;
 import me.manuelp.jevsto.dataTypes.Event;
+import me.manuelp.jevsto.dataTypes.Stream;
 import org.threeten.bp.LocalDateTime;
 import rx.Observable;
 
-import java.util.UUID;
-
+/**
+ * Conceptually, this abstraction represents a dynamic set of {@link Event} streams.
+ * Every stream has a unique {@link Stream identifier} used as key, and can be queried at a specific point in time or
+ * subscribed to (thanks to RxJava).
+ */
 public interface EventStore {
-  void append(Event e);
+  void append(Stream stream, Event e);
 
-  Observable<Event> getEvents();
-
-  Observable<Event> getAllEvents();
-
-  Observable<Event> getAllEventsFrom(LocalDateTime from);
-
-  List<Event> getAll();
-
-  List<Event> getFrom(LocalDateTime t);
+  List<Stream> getStreams();
 
   Option<Event> getById(UUID id);
+
+  Observable<Event> getEvents(Stream stream);
+
+  Observable<Event> getAllEvents(Stream stream);
+
+  Observable<Event> getAllEventsFrom(Stream stream, LocalDateTime from);
+
+  List<Event> getAll(Stream stream);
+
+  List<Event> getFrom(Stream stream, LocalDateTime t);
 }